dodanie package-info.java (wyswietlanie LocalDate w XML) oraz pobieranie danych z API - zczytanie metody

+++ src/main/java/pl/ug/ug_ci/model/ConverterDto.java
@@ -2,15 +2,11 @@ package pl.ug.ug_ci.model;
 
 import lombok.Builder;
 import lombok.Getter;
-
 import java.time.LocalDate;
 
 @Getter
 @Builder
 public class ConverterDto {
-    //TODO: postingDate z ConverterDto i (LocalDate orderPostingDate z Orders) jako jedna zmienna
     private LocalDate postingDate;
-    private Double exchangeRate;
-
-
+    private Float exchangeRate;
 }
